More GUI fixes
Added a tutorial button for both offline games and online games.

Also fixed the slideshow so that it's not longer possible to click next and previous forever. When you reach the final slide the next button becomes disabled and when you reach the first slide the prevButton does the same.

Did some refactoring of the code in the GUI classes as well